requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
WTForms==3.1.0
SQLAlchemy==2.0.23
argon2-cffi==23.1.0
itsdangerous==2.1.2
email-validator==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0